requirements.txt
aiogram==3.11.0
python-dotenv
aiofiles
asyncpg
aiohttp